securethread_backend/requirements.txt
# Core
Django>=5.0.0,<5.3.0
djangorestframework>=3.14.0,<4.0.0
djangorestframework-simplejwt>=5.3.0,<6.0.0
djoser>=2.2.0,<3.0.0
django-environ>=0.11.0,<1.0.0
python-magic>=0.4.27,<0.5.0
pyclamd>=0.4.0,<0.5.0
django-storages>=1.14.2,<2.0.0
boto3>=1.34.0,<2.0.0

# Authentication & Security
django-axes>=6.1.0,<7.0.0
django-cors-headers>=4.3.0,<5.0.0
django-ratelimit>=4.1.0,<5.0.0
django-guardian>=2.4.0,<3.0.0
argon2-cffi>=23.1.0,<24.0.0
django-simple-captcha>=0.6.0,<0.7.0
bleach>=6.1.0,<7.0.0
requests>=2.31.0,<3.0.0
pyotp>=2.9.0,<3.0.0
qrcode>=7.4.2,<8.0.0

# API Documentation
drf-yasg>=1.21.7,<2.0.0
django-filter>=23.5,<24.0.0
markdown>=3.5.1,<4.0.0

# Real-time Features
channels>=4.0.0,<5.0.0
channels-redis>=4.1.0,<5.0.0
redis>=5.0.0,<6.0.0

# Database
psycopg>=3.1.15,<4.0.0

# Testing
pytest>=7.4.3,<8.0.0
pytest-django>=4.7.0,<5.0.0
pytest-asyncio>=0.23.2,<0.24.0
factory-boy>=3.3.0,<4.0.0
coverage>=7.3.4,<8.0.0

# Development
black>=23.11.0,<24.0.0
isort>=5.12.0,<6.0.0
flake8>=6.1.0,<7.0.0
mypy>=1.7.1,<2.0.0

# Production
gunicorn>=21.2.0,<22.0.0
whitenoise>=6.6.0,<7.0.0
Pillow>=10.1.0,<11.0.0 